Fix all 91 SonarQube BUG-type issues on PR 4 (reliability gate)
- S853: explicit casts on <</~ results (flash_log, factory_image, uds,
  uds_ota, uds_log_download, poseidon_accessories, divecan_tx, tests)
- S874: unsigned operands for bitwise ops (poseidon_accessories rc
  accumulators, flash_log, firmware_confirm/divecan_tx tests)
- S867: explicit bool comparisons, no bitwise on bool
- S1226: locals instead of parameter reassignment (incl. solenoid
  arm_timer clamp now correctly fed to counter_us_to_ticks)
- S966/M23_045: #if defined() guards for CONFIG_SOL_O2_INJECT_2_CHANNEL /
  CONFIG_*_TRANSDUCER_CHANNEL — an undefined int Kconfig symbol would
  previously have evaluated 0>=0 and silently enabled the guarded code
- S3519 blockers: explicit clamp in send_uds test helpers
- S1244: float != replaced with memcmp bit-change detection (test shim)
- Web accessibility: labels/aria-labels + keyboard handler (DiveCAN_bt)

Also repairs pre-existing native test breakage from 2fb93e9: uds_ota and
uds_state_did_ota did not link (missing factory_image_restore_async
stub/wrap); F276 tests updated to assert the async restore kick.

+++ Firmware/src/divecan/divecan_tx.c
@@ -120,7 +120,7 @@ void txStatus(DiveCANType_t deviceType, BatteryV_t batteryVoltage,
     uint8_t errByte = (uint8_t)error;
     if (((errByte & DIVECAN_ERR_BAT_MASK) != (uint8_t)DIVECAN_ERR_BAT_LOW)
         && showBattery) {
-        errByte = (uint8_t)((errByte & ~DIVECAN_ERR_BAT_MASK)
+        errByte = (uint8_t)((errByte & (uint8_t)~DIVECAN_ERR_BAT_MASK)
                             | (uint8_t)DIVECAN_ERR_BAT_NORM);
     }
